Added the first stage of project designated as "Mark 1"
Record a live audio sample using the computer's default microphone and
process the sample post-recording.
Command line based program that asks the user's input.
Need to implement a loop to keep the program running to ask the user if
they would like to modulate the same sample or record again.