Added the AudioSpectra file
Added the AudioSpectra file; which was an attempt to do real-time voice
modulation and display the audio on a spectrum analyser and time plot.
There is a time delay of approximately 3 seconds between when the audio
is recorded and when the modulated audio is heard through the speakers.
Assumed to be due to the pipeling effect in the modulation process.